requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fonttools==4.55.8
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
openai==1.61.1
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
